Reading in file's
This code reads in files to both MATPower and MATLAB's built in graph tools. I will be writing more code, this was fairly simple but took me a while to figure out. The coding for the model will take longer, I am already working on that as well. I will likely be writing several functions for our use as well. More to follow...